Move Unix peer credentials tests to their own file
